fix build

+++ include/JSystem/J3DGraphBase/J3DTexture.h
@@ -93,7 +93,7 @@ struct J3DTexCoord : public J3DTexCoordInfo {
 
     u8 getTexGenType() const { return mTexGenType; }
     u8 getTexGenSrc() const { return mTexGenSrc; }
-    u32 getTexGenMtx() const { return mTexGenMtx; }
+    u8 getTexGenMtx() const { return mTexGenMtx; }
     u32 getTexMtxReg() const { return mTexMtxReg & 0xff; }
     void setTexGenMtx(u8 param_1) { mTexGenMtx = param_1; }
     void setTexMtxReg(u16 reg) { mTexMtxReg = reg; }
